Rename `pos` to `source`
