+++ bon-macros/src/util/mod.rs
@@ -28,8 +28,7 @@ pub(crate) mod prelude {
     pub(crate) use super::attrs::AttributeExt;
     pub(crate) use super::fn_arg::FnArgExt;
     pub(crate) use super::ident::IdentExt;
-    pub(crate) use super::iterator::IntoIteratorExt;
-    pub(crate) use super::iterator::IteratorExt;
+    pub(crate) use super::iterator::{IntoIteratorExt, IteratorExt};
     pub(crate) use super::path::PathExt;
     pub(crate) use super::punctuated::PunctuatedExt;
     pub(crate) use super::ty::TypeExt;
